Separate sign up implementation
