+++ src/components/header/header.component.jsx
@@ -7,7 +7,7 @@ import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import { useHistory } from 'react-router-dom';
 
-import { auth } from '../../firebase/firebase.utils';
+import { auth } from '../../firebase/firebase';
 
 import './header.styles.scss';
